refactor: extract useSessionDerived composable from App.vue
9 computed properties moved: activeSession, toolResults,
sidebarResults, currentSummary, isRunning, statusMessage,
toolCallHistory, activeSessionCount, unreadCount.
